Добавляет модули
